refactor(sloppy): give every tool one file, and take templates off the agent
Each tool now carries its description, schemas and executor in one
defineTool file, with registry.ts as the single assembly point; specs.ts
is gone. The apply_template tool goes with it: templates are applied
directly from the composer, and the agent already sees the active one in
its per-request context.
