+++ lib/agent/tools/setCharacter.ts
@@ -1,13 +1,31 @@
-import type { AgentToolContext } from "./context";
-import type { ToolInput } from "./specs";
+import dedent from "dedent";
+import { z } from "zod";
+import { defineTool } from "./defineTool";
+import { VOICE_TRAITS } from "./inputs";
 
-export async function setCharacter(
-	{ name, ...patch }: ToolInput<"set_character">,
-	ctx: AgentToolContext,
-): Promise<string> {
-	const settled = ctx.setCharacter(name, patch);
+export const setCharacter = defineTool({
+	description: dedent`
+	  Set what a character looks like or sounds like, creating it if the name is new.
+	  Reference a character by the exact \`name\` its lines use in the script.
 
-	return settled.created
-		? `Added ${settled.name}. Its avatar has not been generated yet.`
-		: `Changed ${settled.name}.`;
-}
+	  Send only what changes. \`appearance\` is what the character's avatar is drawn from, and
+	  is always English. The rest describe the voice it speaks in.
+	`,
+	input: z.object({
+		name: z.string().min(1),
+		appearance: z
+			.string()
+			.min(1)
+			.optional()
+			.describe("What the character looks like, in English."),
+		...VOICE_TRAITS,
+	}),
+	output: z.string(),
+	execute: async ({ name, ...patch }, ctx) => {
+		const settled = ctx.setCharacter(name, patch);
+
+		return settled.created
+			? `Added ${settled.name}. Its avatar has not been generated yet.`
+			: `Changed ${settled.name}.`;
+	},
+});
