+++ lib/agent/tools/setLanguage.ts
@@ -1,11 +1,21 @@
-import { languageLabel } from "@/lib/project/language";
-import type { AgentToolContext } from "./context";
-import type { ToolInput } from "./specs";
+import dedent from "dedent";
+import { z } from "zod";
+import { LANGUAGE_CHOICES, languageLabel } from "@/lib/project/language";
+import { defineTool } from "./defineTool";
 
-export async function setLanguage(
-	{ language }: ToolInput<"set_language">,
-	ctx: AgentToolContext,
-): Promise<string> {
-	ctx.setMetadata({ language });
-	return `Set the language to ${languageLabel(language)}. It applies to the next script written; what is on the canvas is unchanged.`;
-}
+export const setLanguage = defineTool({
+	description: dedent`
+	  Set the language narration and dialogue are written in. "auto" follows whatever language
+	  the user writes in.
+
+	  This applies to scripts written from here on. It does not translate what is on the canvas.
+
+	  Languages: ${LANGUAGE_CHOICES.map((c) => `${c} (${languageLabel(c)})`).join(", ")}
+	`,
+	input: z.object({ language: z.enum(LANGUAGE_CHOICES) }),
+	output: z.string(),
+	execute: async ({ language }, ctx) => {
+		ctx.setMetadata({ language });
+		return `Set the language to ${languageLabel(language)}. It applies to the next script written; what is on the canvas is unchanged.`;
+	},
+});
